backend/requirements.txt

pytesseract==0.3.10
opencv-python-headless==4.10.0.84
Pillow==10.4.0
python-multipart==0.0.9
fastapi==0.115.0
uvicorn[standard]==0.30.6
